+++ cgo/kuzzle/errors.go
