--- cgo/kuzzle/get_my_credentials.go
@@ -1,35 +1,0 @@
-package main
-
-/*
-	#cgo CFLAGS: -I../../headers
-	#cgo LDFLAGS: -ljson-c
-	#include <stdlib.h>
-	#include "kuzzle.h"
-*/
-import "C"
-import (
-	"encoding/json"
-	"github.com/kuzzleio/sdk-go/kuzzle"
-	"unsafe"
-)
-
-//export kuzzle_wrapper_get_my_credentials
-func kuzzle_wrapper_get_my_credentials(k *C.kuzzle, strategy *C.char, options *C.query_options) *C.json_result {
-	result := (*C.json_result)(C.calloc(1, C.sizeof_json_result))
-	opts := SetQueryOptions(options)
-
-	res, err := (*kuzzle.Kuzzle)(k.instance).GetMyCredentials(C.GoString(strategy), opts)
-
-	if err != nil {
-		Set_json_result_error(result, err)
-		return result
-	}
-
-	r, _ := json.Marshal(res)
-	buffer := C.CString(string(r))
-	defer C.free(unsafe.Pointer(buffer))
-
-	result.result = C.json_tokener_parse(buffer)
-
-	return result
-}
