+++ cgo/kuzzle/kuzzle.go
@@ -7,6 +7,10 @@ package main
 	#include "kuzzle.h"
 
 	typedef query_object *query_object_ptr;
+
+	static void call(void* f, json_object* res) {
+		((void(*)(json_object*))f)(res);
+	}
 */
 import "C"
 import (
@@ -16,6 +20,14 @@ import (
 	"github.com/kuzzleio/sdk-go/kuzzle"
 	"unsafe"
 )
+
+// map which stores instances to keep references in case the gc passes
+var instances map[interface{}]interface{}
+
+// register new instance to the instances map
+func registerKuzzle(instance interface{}) {
+	instances[instance] = nil
+}
 
 // unregister an instance from the instances map
 //export unregisterKuzzle
@@ -38,7 +50,7 @@ func kuzzle_wrapper_new_kuzzle(k *C.kuzzle, host, protocol *C.char, options *C.o
 	}
 
 	inst, _ := kuzzle.NewKuzzle(c, opts)
-	registerInstance(inst)
+	registerKuzzle(inst)
 
 	k.instance = unsafe.Pointer(inst)
 }
@@ -52,6 +64,21 @@ func kuzzle_wrapper_connect(k *C.kuzzle) *C.char {
 	}
 
 	return nil
+}
+
+//export kuzzle_wrapper_disconnect
+func kuzzle_wrapper_disconnect(k *C.kuzzle) {
+	(*kuzzle.Kuzzle)(k.instance).Disconnect()
+}
+
+//export kuzzle_wrapper_set_default_index
+func kuzzle_wrapper_set_default_index(k *C.kuzzle, index *C.char) C.int {
+	err := (*kuzzle.Kuzzle)(k.instance).SetDefaultIndex(C.GoString(index))
+	if err != nil {
+		return C.int(C.EINVAL)
+	}
+
+	return 0
 }
 
 //export kuzzle_wrapper_get_offline_queue
@@ -81,10 +108,68 @@ func kuzzle_wrapper_get_offline_queue(k *C.kuzzle) *C.offline_queue {
 	return result
 }
 
-// Allocates memory
-//export kuzzle_wrapper_get_jwt
-func kuzzle_wrapper_get_jwt(k *C.kuzzle) *C.char {
-	return C.CString((*kuzzle.Kuzzle)(k.instance).GetJwt())
+//export kuzzle_wrapper_flush_queue
+func kuzzle_wrapper_flush_queue(k *C.kuzzle) {
+	(*kuzzle.Kuzzle)(k.instance).FlushQueue()
+}
+
+//export kuzzle_wrapper_replay_queue
+func kuzzle_wrapper_replay_queue(k *C.kuzzle) {
+	(*kuzzle.Kuzzle)(k.instance).ReplayQueue()
+}
+
+//export kuzzle_wrapper_start_queuing
+func kuzzle_wrapper_start_queuing(k *C.kuzzle) {
+	(*kuzzle.Kuzzle)(k.instance).StartQueuing()
+}
+
+//export kuzzle_wrapper_stop_queuing
+func kuzzle_wrapper_stop_queuing(k *C.kuzzle) {
+	(*kuzzle.Kuzzle)(k.instance).StopQueuing()
+}
+
+//export kuzzle_wrapper_get_headers
+func kuzzle_wrapper_get_headers(k *C.kuzzle) *C.json_object {
+	res := (*kuzzle.Kuzzle)(k.instance).GetHeaders()
+	r, _ := json.Marshal(res)
+
+  buffer := C.CString(string(r))
+  defer C.free(unsafe.Pointer(buffer))
+	
+  return C.json_tokener_parse(buffer)
+}
+
+//export kuzzle_wrapper_set_headers
+func kuzzle_wrapper_set_headers(k *C.kuzzle, content *C.json_object, replace C.uint) {
+	if JsonCType(content) == C.json_type_object {
+		r := replace != 0
+		(*kuzzle.Kuzzle)(k.instance).SetHeaders(JsonCConvert(content).(map[string]interface{}), r)
+	}
+}
+
+//export kuzzle_wrapper_add_listener
+func kuzzle_wrapper_add_listener(k *C.kuzzle, e C.int, cb unsafe.Pointer) {
+	c := make(chan interface{})
+
+	kuzzle.AddListener((*kuzzle.Kuzzle)(k.instance), int(e), c)
+	go func() {
+		res := <-c
+
+		var jsonRes *C.json_object
+		r, _ := json.Marshal(res)
+
+		buffer := C.CString(string(r))
+		defer C.free(unsafe.Pointer(buffer))
+
+		jsonRes = C.json_tokener_parse(buffer)
+
+		C.call(cb, jsonRes)
+	}()
+}
+
+//export kuzzle_wrapper_remove_listener
+func kuzzle_wrapper_remove_listener(k *C.kuzzle, event C.int) {
+	(*kuzzle.Kuzzle)(k.instance).RemoveListener(int(event))
 }
 
 func main() {
