--- cgo/kuzzle/remove_listener.go
@@ -1,13 +1,0 @@
-package main
-
-/*
-	#cgo CFLAGS: -I../../headers
-	#include "kuzzle.h"
-*/
-import "C"
-import "github.com/kuzzleio/sdk-go/kuzzle"
-
-//export kuzzle_wrapper_remove_listener
-func kuzzle_wrapper_remove_listener(k *C.kuzzle, event C.int) {
-	(*kuzzle.Kuzzle)(k.instance).RemoveListener(int(event))
-}
