--- cgo/kuzzle/replay_queue.go
@@ -1,13 +1,0 @@
-package main
-
-/*
-	#cgo CFLAGS: -I../../headers
-	#include "kuzzle.h"
-*/
-import "C"
-import "github.com/kuzzleio/sdk-go/kuzzle"
-
-//export kuzzle_wrapper_replay_queue
-func kuzzle_wrapper_replay_queue(k *C.kuzzle) {
-	(*kuzzle.Kuzzle)(k.instance).ReplayQueue()
-}
